requirements.txt
Flask>=2.0
pytest
requests
python-dotenv
